refactor: sending of payload identifier to services
