+++ utils/getFullPrice.ts
